Added word definition js and csv-to-keyed-json.py
Added functionality in the "selectWord" function, that fetches the
definition and grammar explanation of a latin word, from the
'words.json' file, and outputs it to the user.

I also added a script called 'csv-to-keyed-json.py,' that utilised the
'pandas' python library, to read the 'words.csv' file, sort it in
alphabetical order, and output the result as a keyed json file.